Correctly split into even and odd parts
I am quite positive this could be done in the main FFT loop, and thus not require this level of array munging during setup. However, this now passes all of the test cases, so we'll worry about speed later.

Refactor fft_test a bit and add some test cases that were broken before.